docs: restructure example with multi-module and go.work
- Split into basic/ (struct + markers) and basic-usage/ (demo code)
- Add go.work for workspace configuration
- Users can now regenerate validation code safely
- Remove go.work from .gitignore
