+++ example/basic-usage/main.go
@@ -5,18 +5,19 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/sivchari/govalid/example/basic"
 	govaliderrors "github.com/sivchari/govalid/validation/errors"
 )
 
 func main() {
 	// Example 1: Valid user
 	fmt.Println("=== Example 1: Valid User ===")
-	validUser := &User{
+	validUser := &basic.User{
 		Name:  "John Doe",
 		Email: "john@example.com",
 		Age:   30,
 	}
-	if err := ValidateUser(validUser); err != nil {
+	if err := basic.ValidateUser(validUser); err != nil {
 		fmt.Printf("Validation failed: %v\n", err)
 	} else {
 		fmt.Println("Validation passed!")
@@ -24,18 +25,18 @@ func main() {
 
 	// Example 2: Invalid user with multiple errors
 	fmt.Println("\n=== Example 2: Invalid User ===")
-	invalidUser := &User{
+	invalidUser := &basic.User{
 		Name:  "",              // required, minlength=1 violation
 		Email: "invalid-email", // email format violation
 		Age:   200,             // lte=150 violation
 	}
-	if err := ValidateUser(invalidUser); err != nil {
+	if err := basic.ValidateUser(invalidUser); err != nil {
 		fmt.Printf("Validation failed: %v\n", err)
 	}
 
 	// Example 3: Loop through all validation errors
 	fmt.Println("\n=== Example 3: Loop Through Errors ===")
-	if err := ValidateUser(invalidUser); err != nil {
+	if err := basic.ValidateUser(invalidUser); err != nil {
 		// Type assert to ValidationErrors to iterate over individual errors
 		if errs, ok := err.(govaliderrors.ValidationErrors); ok {
 			fmt.Printf("Found %d validation errors:\n", len(errs))
@@ -50,7 +51,7 @@ func main() {
 
 	// Example 4: Check for specific error type
 	fmt.Println("\n=== Example 4: Check Specific Error ===")
-	if err := ValidateUser(invalidUser); err != nil {
+	if err := basic.ValidateUser(invalidUser); err != nil {
 		if errs, ok := err.(govaliderrors.ValidationErrors); ok {
 			for _, e := range errs {
 				if e.Type == "required" {
