+++ example/go.work
@@ -1,0 +1,6 @@
+go 1.24.0
+
+use (
+	./basic
+	./basic-usage
+)
